chore: add ESLint configuration file with particular rules
Added a new ESLint configuration file. This configuration includes TypeScript and unicorn ESLint plugin to help with the static analysis of the codebase. It also aids in identifying problematic patterns found in code. The 'unicorn/prevent-abbreviations' rule has been specifically turned off to avoid unnecessary flags on abbreviations in code. This provides a standardized coding approach across the project.