ci: configure dependabot for project
The addition of the dependabot configuration file allows for automated updates of dependencies. This change enables the project to stay up-to-date with the latest versions of dependencies for both GitHub Actions and PNPM every day. It also maintains a pull request limit of 20 for GitHub Actions dependencies and 10 for PNPM dependencies. This automated process ensures regular, manageable updates, thus improving efficiency and potential security risks.