chore: enable kv namespace in worker configuration
The KV namespace in the worker configuration is enabled in order to use KV as persistent storage for small key-value pairs. This was done by uncommenting the kv_namespaces block and providing the binding and id in the wrangler.toml file. Additionally, the generated comments in worker-configuration.d.ts file have been updated to reflect the running of the newer `wrangler types` command.