refactor: update scheduled function and remove unnecessary comments
The scheduled function in 'src/index.ts' was updated to handle fetching accounts. This involved removing the current code functionality and replacing it with fetching accounts from 'services/lilnouns/fetch-accounts' if null. The fetched accounts are then stored in KV with an expiration time to live of 24 hours. This change focuses on required functionality enhancement. Furthermore, extensive comments about Workers that initially existed in the code have been removed. This is done to declutter the code and facilitates readability by focusing only on necessary information.